Scraper/requirements.txt
playwright>=1.40.0
crawl4ai>=0.3.0
google-generativeai>=0.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
asyncio
